Fixes enchantment emulation in 1.20.5->1.21 on non paper servers
Moves out Paper specific code so the fix can also run on non-Paper server and also uses older/deprecated API to run on legacy versions like 1.8-1.12.
Note that this still doesn't support most enchantments on older/non-Paper servers, but the most important enchantment is supported now.

Closes https://github.com/ViaVersion/ViaVersion/issues/3951
