+++ bukkit/src/main/java/com/viaversion/viaversion/bukkit/platform/BukkitViaLoader.java
@@ -33,6 +33,7 @@ import com.viaversion.viaversion.bukkit.listeners.protocol1_9to1_8.BlockListener
 import com.viaversion.viaversion.bukkit.listeners.protocol1_9to1_8.DeathListener;
 import com.viaversion.viaversion.bukkit.listeners.protocol1_9to1_8.HandItemCache;
 import com.viaversion.viaversion.bukkit.listeners.protocol1_9to1_8.PaperPatch;
+import com.viaversion.viaversion.bukkit.listeners.v1_20_5to1_21.PaperPlayerChangeItemListener;
 import com.viaversion.viaversion.bukkit.listeners.v1_20_5to1_21.PlayerChangeItemListener;
 import com.viaversion.viaversion.bukkit.providers.BukkitAckSequenceProvider;
 import com.viaversion.viaversion.bukkit.providers.BukkitBlockConnectionProvider;
@@ -179,8 +180,12 @@ public class BukkitViaLoader implements ViaPlatformLoader {
             Via.getManager().getProviders().use(AckSequenceProvider.class, new BukkitAckSequenceProvider(plugin));
             new BlockBreakListener(plugin).register();
         }
-        if (serverProtocolVersion.olderThan(ProtocolVersion.v1_21) && PaperViaInjector.hasClass("io.papermc.paper.event.player.PlayerInventorySlotChangeEvent")) {
-            new PlayerChangeItemListener(plugin).register();
+        if (serverProtocolVersion.olderThan(ProtocolVersion.v1_21)) {
+            if (PaperViaInjector.hasClass("io.papermc.paper.event.player.PlayerInventorySlotChangeEvent")) {
+                new PaperPlayerChangeItemListener(plugin).register();
+            } else {
+                new PlayerChangeItemListener(plugin).register();
+            }
         }
     }
